api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
PyJWT==2.8.0
cryptography>=3.4.8
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.3
psycopg2-binary==2.9.9
PyPDF2==3.0.1
requests==2.31.0
python-dotenv==1.0.0
usaddress==0.5.11
reportlab==4.0.7
